mm/migrate_pages: split unmap_and_move() to _unmap() and _move()
This is a preparation patch to batch the page unmapping and moving
for the normal pages (including THP).  Based on that we can batch the
TLB shootdown during the page migration and use some hardware
accelerator for the page copying.

In this patch, unmap_and_move() is split to migrate_page_unmap() and
migrate_page_move().  To pass some information between unmap and move,
the original unused newpage->mapping and newpage->private is used.

Signed-off-by: "Huang, Ying" <ying.huang@intel.com>