x86/sgx: Save enclave pointer for VA page
An enclave VA page has struct sgx_va_page as its *owner*, which can
tell if the VA page was initialized for an enclave. However, if
the VA page has failed to be EREMOVEd while enclave is exiting, the
*owner* field will not be cleared even if the enclave is destroyed.
Consequently, it's hard to say if the VA page is still associated
with an enclave.

Define an enclave pointer field in struct sgx_va_page, and save
the actual value when the VA page is initialized for an enclave.
Once the enclave is destroyed, the enclave pointer will become
invalid, so, it can tell exactly if the VA page is associated
with an enclave.

This information is useful for SGX CPUSVN update process, which
will rely on it to decide how to zap the VA page.

Note that the new pointer defined for struct sgx_va_page will
mostly consumed 8M on a system with 2T of SGX EPC memory configured
only when all the SGX EPC memory are allocated for enclaves
concurrently running on host. For it's a rarely scenario, the actual
memory consumption should be less than the estimated value.

Signed-off-by: Cathy Zhang <cathy.zhang@intel.com>