x86/microcode: Add an option to reload microcode even if revision is
unchanged

From: Ashok Raj <ashok.raj@intel.com>

This is POC to support rollback. This is a simple version, admin uses
echo 2 instead of echo 1 to reload. We don't do the version checks.

The following usage, writing 2 to reload file is helpful to reload
the microcode again even if the revision is less than what is loaded.

Signed-off-by: Ashok Raj <ashok.raj@intel.com>