x86/cpu/tdx: Load the TDX firmware as the second step
Loading the TDX module firmware requires a signature structure to verify
the TDX module has not been maliciously tampered.  Loading TDX module
requires to call P-SEAMLDR SEAMCALL on all logical CPUs on the platform,
passing the firmware module and signature structure as input.

If the necessary CPU feature, CPU SEAM mode, is available, try to find the
TDX module and its signature file in the built-in firmware or initrd.  If
they're both found, load the TDX module into the isolated memory range.

Co-developed-by: Sean Christopherson <sean.j.christopherson@intel.com>
Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>
Co-developed-by: Kai Huang <kai.huang@intel.com>
Signed-off-by: Kai Huang <kai.huang@intel.com>
Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>