PCI/MSI: Add mutex to serialize MSI-X interrupt allocation
This is a preparatory patch to enable dynamic allocation of MSI-X
interrupts.

After enabling dynamic allocation of MSI-X interrupts, prevent
concurrent MSI-X interrupt allocations to the same device by using a
mutex. Until now, the device msi_list was immutable but once dynamic
allocation of MSI-X is enabled, entries can be added or removed from
this list, and thus needs to be protected.

No functional change.

Signed-off-by: Megha Dey <megha.dey@intel.com>