x86/cpu/tdx: Shrink number of TDMR ranges when necessary
Kernel builds a list of converted ranges, or TDMR ranges, to convert all
TDX memory blocks as TDX memory.  When there are many TDX memory blocks,
there might be many TDMR ranges as well.  Final TDMRs are constructed based
on the TDMR range, which means the number of final TDMRs cannot be smaller
than the number of TDMR ranges.  TDX only supports a limited number of
TDMRs, so when there are too many TDMR ranges, it's not possible to
generate a supported number of TDMRs.

Normally, system memory from memblock won't have too many discrete blocks,
but on a large system, or with kernel parameter 'memmap=nn!ss' (or
'memmap=...' in general), it could end up with many discrete TDX memory
blocks in final TDX memory.  Therefore, although rare, it's possible that
the number of TDMR ranges could be larger than TDX can support (currently
64).

To allow to continue to construct TDMRs when there are too many TDMR
ranges, shrink the number of TDMR ranges by merging adjacent ranges,
until it is not bigger than the number of TDMRs that TDX can support.

Signed-off-by: Kai Huang <kai.huang@intel.com>
Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>