x86/tdx: Basic TDCALL fuzzing support
In TDX the hypervisor is untrusted and can attack the kernel
through the output of MMIO/MSR reads/port IO and some other options.
Untrusted "hardware" is a need concept in Linux, so we need to make sure the
code is hardened against bad responses.

Add a simple fuzzer to allow hardening the interfaces. This is not
sufficient for full hardening, but useful for basic testing.

This hooks into the low level TDcalls and allows to fuzz all
interactions with the host, except those through shared memory.

The basic setup is through the fault injection framework, but
in addition there is own random generation state that allows
flipping random bits in the TDCALL responses.

/sys/kernel/debug/tdx_fail/ Tunables
(in addition to standard fault injection tunables like skip etc.):
seed: Set the seed for the random generator per CPU.
tdcall: Enable bit flipping for TDCALLs
num_change_bits: How many bits to flip in a TDCALL response.
		 Set to 64 to replace the complete value.

The same can be also configure with a command line parameter
for early  boot coverage

Signed-off-by: Andi Kleen <ak@linux.intel.com>