dmaengine: idxd: add support for inter-pasid table window attach ioctl
The window attach ioctl is used after the window create ioctl. It is used
by a submitter process to attach to the IDPT entry file handle. The
submitter's pasid is enabled for the IDPT entry. At this point, the driver
sets the IDPT entry's valid flag to 1 to enable the entry for the device.
The ioctl returns an IDPT entry handle that is needed by the inter-pasid
DMA descriptor submitted by the submitter.

Signed-off-by: Dave Jiang <dave.jiang@intel.com>