x86/cpu/tdx: Launch the TDX first firmware loader
Launch the first firmware loader(NP-SEAMLDR) by calling the helper function
that preserves CPU status as a normal C function.

Because The NP-SEAMLDR requires application processors(APs) in a special
state, Wait-For-SIPI, and entropy to generate random numbers.  Put
application processors(APs) in a special state before launching the
NP-SEAMLDR.  If APs aren't in the required state or entropy is lacking,
wait for a while and retry because there is no way to detect if AP is in
Wait-For-SIPI state or to detect if there is enough entropy.

Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>