asm/io.h: Add ioremap_driver_hardening
This function is for declaring memory that should be shared with
a hypervisor in a confidential guest. If the architecture doesn't
implement it it's just ioremap.

Note this function is not supposed to be used directly by drivers.
But it will be used implicitly by pci_iomap -- based on the
authorized attribute -- and also by the PCI subsystem. To avoid
ifdefs we create a generic ioremap alias.

Signed-off-by: Andi Kleen <ak@linux.intel.com>
Signed-off-by: Kuppuswamy Sathyanarayanan <sathyanarayanan.kuppuswamy@linux.intel.com>