x86/fpu: Prepare KVM for updating XFD state
When XFD is non-trapped, it should be ensured that XFD and
guest_fp::fpstate::xfd are consistency at exit of fastpath loop to avoid
breaking the existing fpstate mechanism. Because within the fastpath loop,
the preemption and interrupt are both disabled so there is no chance for
KVM or native to use or swap the XFD.

Provide the function for KVM. So when XFD passthrough, at exit of
fastpath loop, KVM uses to update the state accordingly.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Jing Liu <jing2.liu@intel.com>