platform/x86: intel_tdx_attest: change tdx_attest_ioctl TDX_CMD_GEN_QUOTE ioctl interface
The input parameter of TDX_CMD_GEN_QUOTE ioctl is changed to the struct
tdx_gen_quote, which contains the user pointer of a message buffer and its
buffer size. Copy the message buffer as the parameter of
tdvmcall[GetQuote] to host VMM.

Upon receiving the notify event, the overridden data is already saved in
shared message buffer. Copy it back to userspace.

Signed-off-by: Chenyi Qiang <chenyi.qiang@intel.com>