x86/virtext: Split out VMX enabling code from kvm-intel
Both TDX module driver (code that initializes TDX module) and kvm-intel
rely on VMX. Move VMX enabling code out of kvm-intel to a common place
for reuse.

Since multiple components can use VMX, disabling VMX when anyone is using
it would lead to unexpected failures. So, a refcount is introduced to track
the number of references of VMX on each CPU along with two APIs to
in(de)crease it.

Note that
1. KVM performs compatibility checks before enabling VMX on any CPU. This
common code doesn't do such checks, then VMX is allowed to be enabled
on some CPUs even if there are other CPUs that don't support VMX. But an
error is returned if trying to enable VMX on CPUs lack of VMX capability.

2. VMX reference count is tracked in per-cpu scope. This common code
doesn't ensure consistent VMX state (enabled or disabled) system-wide.
Users of VMX that need consistent VMX state across all CPUs should enforce
it by themselves.

3. intel_pt_handle_vmx() is used to prevent host kernel from enabling PT
when VMX is enabled because initial implementations of Intel PT do not
support tracing in VMX operation. It is just noop for later
implementations.

Signed-off-by: Chao Gao <chao.gao@intel.com>