memremap_pages: Introduce devmap_protected()
Users of protected dev_pagemaps can check the PGMAP_PROTECTION flag to
see if the devmap is protected.  However, most callers operate on struct
page's not the pagemap directly.

Define devmap_protected() to determine if a page is part of a
dev_pagemap mapping and if so if the page is protected by the additional
protections.

Signed-off-by: Ira Weiny <ira.weiny@intel.com>