KVM: x86/legacy mmu: prohibit mapping private or shared page for legacy MMU
SPTE_SHARED_MASK in non-present private spte indicates that the GPA is
allowed to map as shared (not-allowed to map as private) if set.  (or
private if cleared).

On kvm page fault on private GPA, check if the bit is set in the
non-present leaf spte.  If the bit is set, don't resolve the fault and let
vcpu keep faulting (until other vcpu changes the bit). On kvm page fault on
shared GPA, check if the bit is set in corresponding private leaf spte.  If
cleared (or leaf spte isn't allocated yet), don't resolve kvm page fault.

Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>