crypto: iax - Add iax_crypto_enable module param
Because neither iax_crypto nor the target device operating on top of
it can determine whether or the user setup is complete e.g. all the
required iax wqs have been completely configured and initialized, a
separate param is required to ensure that compression/decompression
only happens when ready.

In order to actually start/stop compression/decompression, the
iax_crypto_enable module param should be used.

To start compression/decompression, write a '1' to the file:

  echo 1 > /sys/module/iax_crypto/parameters/iax_crypto_enable

To stop compression/decompression, write a '0' to the file:

  echo 0 > /sys/module/iax_crypto/parameters/iax_crypto_enable

Signed-off-by: Tom Zanussi <tom.zanussi@intel.com>