mm/pagezero: Add page clear engine driver
This driver makes use of a DMA device through the dmaengine API to
allocate and clear kernel memory to build a pool of pre-zeroed 4KB pages
on each NUMA node to speed up __GFP_ZERO allocations. User must manually
configure the DMA device on each NUMA node before loading this driver.

The driver finds a DMA channel with DMA_MEMSET capability on each NUMA
node, uses the DMA device to clear large pages, and hands out the
pieces.

DMA engine APIs are called to prepare and submit DMA descriptors, and to
check completion status. The dst_addr of descriptor is filled with the
DMA mapped address of the page to be cleared.

When a new large page is handed to the driver, the driver asks the DMA
device to clear it and leave it hanging on the pernode structure for
completion checking later.

The code that hands out small pages also has to check for completion and
adds newly cleared large pages to the list.

The driver keeps an index into the large pages and free from high to
low, but it pushes them onto the head of the list so that consumers will
get pages in low to high sequence.

Cleanup code on module exit waits for in-progress DMA operations and
frees all the pages attached to the perzone structure.

Module parameters "engine_{order|low_water|flag_cc}" are added for
performance tuning.

Signed-off-by: Tony Luck <tony.luck@intel.com>
Co-developed-by: Xiaochen Shen <xiaochen.shen@intel.com>
Signed-off-by: Xiaochen Shen <xiaochen.shen@intel.com>