perf/x86/intel: Support auto counter reload (POC)
The relative rates among two or more events are useful for performance
analysis, e.g., a lower IPC may indicates an performance issue.
Traditional sampling takes samples of events separately. To get the
relative rates among two or more events, high sample rate is required,
which can brings high overhead. Usually, the relative rates which exceed
some threshold is more useful. The traditional way takes samples
unconditionally. Many samples taken in the non-hotspot area may be
dropped in the post process.

The auto counter reload feature take samples when the relative rate
of two or more events exceeds some threshold, which provide the
fine-grained information with low cost.

The auto counter reload feature is enumerated in CPUID. Add two
variables in the X86_pmu to store the number and mask of the supported
counters.

Add reload in the struct perf_event_attr to indicate that an event
requires the auto counter reload.

All events require the auto counter reload must be in a group. The
leader event must be a PEBS event. It does sampling. The rest of the
member events does counting. All events must have a fixed period.

For example, an user want take samples when IPC <1 across 10K cycle
window. The user can set cycles:p as a leader event, and set
instructions as a member event. Both events enable the auto counter
reload feature and set period to 10K.

If the instructions overflows first, which means IPC > 1, HW reloads
both counters to 10K, no NMI and samples.
If the cycles:p overflows first, a NMI will be triggered, and a sample
is generated. HW reloads the counter of the instructions. NMI handler
reloads the counter of the cycles:p.

The coherent_counter_mask is introduced to store the counters which
are reloaded together. It's maintained by the leader event.

Not all the counters support the auto counter reload feature. Update
the event constraint of the reload event.

TODO:
- Optimize
  Current patch accesses the CFG MSR when enable/disable events.
  May track the value of the CFG MSR, only write it when it's changed.
- If the platform doesn't support the feature, it return -EINVAL as
  well. Need a unique error code for the unsupport case?

Test:
- Need do some tests for the simultaneous Reload
- Need do some tests with half auto reload event and half normal sample
  read.
- Need develop caliborator test case to check e.g., IPC < 1

Signed-off-by: Kan Liang <kan.liang@intel.com>