platform-msi: Add platform check for subdevice irq domain
The pci_subdevice_msi_create_irq_domain() should fail if the
underlying platform is not able to support IMS (Interrupt Message Storage).
Otherwise, the isolation of interrupt is not guaranteed.

For x86, IMS is only supported on bare metal for now. We could
enable it in the virtualization environments in the future if interrupt
HYPERCALL domain is supported or the hardware has the capability of interrupt
isolation for subdevices.

Link:
https://lore.kernel.org/linux-pci/87pn4nk7nn.fsf@nanos.tec.linutronix.de/
Link:
https://lore.kernel.org/linux-pci/877dqrnzr3.fsf@nanos.tec.linutronix.de/
Link:
https://lore.kernel.org/linux-pci/877dqqmc2h.fsf@nanos.tec.linutronix.de/

Signed-off-by: Megha Dey <megha.dey@intel.com>