cxl/pmem: Convert nvdimm bridge API to use dev
The cxl_pmem driver specific cxl_nvdimm structure isn't a suitable
parameter for an exported API that can be used by other drivers.
Instead, use a dev structure, which should be woven into any caller
using this API. This will allow for either the nvdimm's dev, or the
memdev's dev to be used.

Signed-off-by: Ben Widawsky <ben.widawsky@intel.com>
---
Changes since v2:
- Added kdoc to cxl_find_nvdimm_bridge()