thunderbolt: Split setting link width and lane bonding into own functions
XDomain lane bonding, when UUID is "higher", it is the other side of the
link that triggers the lanes to bond. For this reason make setting width
and triggering the actual bonding separate functions that can be called
as needed.

While there remove duplicated empty line in the kernel-doc comment of
tb_port_lane_bonding_disable().

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>